Update the API to be split in admin and normal modes

+++ server/controllers/api/ApiAccountsController.php
@@ -6,10 +6,10 @@ class ApiAccountsController {
         // The pagination vars
         $page = get_page();
         $limit = get_limit();
-        $count = Accounts::count();
+        $count = Accounts::countByUser(Auth::id());
 
         // Select all the accounts by page
-        $accounts = Accounts::selectPage($page, $limit)->fetchAll();
+        $accounts = Accounts::selectPageByUser(Auth::id(), $page, $limit)->fetchAll();
 
         // Return the data as JSON
         return [
@@ -27,10 +27,10 @@ class ApiAccountsController {
         // The pagination vars
         $page = get_page();
         $limit = get_limit();
-        $count = Accounts::searchCount($q);
+        $count = Accounts::searchCountByUser(Auth::id(), $q);
 
         // Select all the accounts by page
-        $accounts = Accounts::searchPage($q, $page, $limit)->fetchAll();
+        $accounts = Accounts::searchSelectPageByUser(Auth::id(), $q, $page, $limit)->fetchAll();
 
         // Return the data as JSON
         return [
@@ -46,17 +46,15 @@ class ApiAccountsController {
         // Validate the user input
         api_validate([
             'name' => Accounts::NAME_VALIDATION,
-            'type' => Accounts::TYPE_VALIDATION,
-            'user_id' => Accounts::USER_ID_VALIDATION,
-            'amount' => Accounts::AMOUNT_VALIDATION
+            'type' => Accounts::TYPE_VALIDATION
         ]);
 
         // Insert the account to the database
         Accounts::insert([
             'name' => request('name'),
             'type' => request('type'),
-            'user_id' => request('user_id'),
-            'amount' => parse_money_number(request('amount'))
+            'user_id' => Auth::id(),
+            'amount' => 0
         ]);
 
         // Return a confirmation message
@@ -68,41 +66,67 @@ class ApiAccountsController {
 
     // The API accounts show route
     public static function show ($account) {
-        return $account;
+        // Check if the account is from the authed user
+        if ($account->user_id == Auth::id()) {
+            return $account;
+        }
+
+        // Return a error message
+        else {
+            return [
+                'message' => 'The account is not yours'
+            ];
+        }
     }
 
     // The API accounts edit route
     public static function edit ($account) {
-        // Validate the user input
-        api_validate([
-            'name' => Accounts::NAME_VALIDATION,
-            'type' => Accounts::TYPE_VALIDATION,
-            'user_id' => Accounts::USER_ID_VALIDATION,
-            'amount' => Accounts::AMOUNT_VALIDATION,
-        ]);
+        // Check if the account is from the authed user
+        if ($account->user_id == Auth::id()) {
+            // Validate the user input
+            api_validate([
+                'name' => Accounts::NAME_VALIDATION,
+                'type' => Accounts::TYPE_VALIDATION
+            ]);
 
-        // Update the account in the database
-        Accounts::update($account->id, [
-            'name' => request('name'),
-            'type' => request('type'),
-            'user_id' => request('user_id'),
-            'amount' => parse_money_number(request('amount'))
-        ]);
+            // Update the account in the database
+            Accounts::update($account->id, [
+                'name' => request('name'),
+                'type' => request('type')
+            ]);
 
-        // Return a confirmation message
-        return [
-            'message' => 'The account has been edited successfully'
-        ];
+            // Return a confirmation message
+            return [
+                'message' => 'The account has been edited successfully'
+            ];
+        }
+
+        // Return a error message
+        else {
+            return [
+                'message' => 'The account is not yours'
+            ];
+        }
     }
 
     // The API accounts delete route
     public static function delete ($account) {
-        // Delete the account
-        Accounts::delete($user->id);
+        // Check if the account is from the authed user
+        if ($account->user_id == Auth::id()) {
+            // Delete the account
+            Accounts::delete($user->id);
 
-        // Return a confirmation message
-        return [
-            'message' => 'The account has been deleted successfully'
-        ];
+            // Return a confirmation message
+            return [
+                'message' => 'The account has been deleted successfully'
+            ];
+        }
+
+        // Return a error message
+        else {
+            return [
+                'message' => 'The account is not yours'
+            ];
+        }
     }
 }
